Proved that the "CodableFeedStore" side-effects run serially
